extract fractions of seconds when extracting thumbnails for videos This addresses bug #305328. The problem was that several frames extracted was the same, and therefore the algorithm for finding the current image was broken, as when I was at image 2, it found image 1 as being the current, as image 1 and 2 both where the same

+++ ImageManager/ExtractOneVideoFrame.cpp
@@ -32,12 +32,12 @@
 namespace ImageManager {
 
 #define STR(x) QString::fromUtf8(x)
-void ExtractOneVideoFrame::extract(const DB::FileName &fileName, int offset, QObject* receiver, const char* slot)
+void ExtractOneVideoFrame::extract(const DB::FileName &fileName, double offset, QObject* receiver, const char* slot)
 {
     new ExtractOneVideoFrame(fileName, offset, receiver, slot);
 }
 
-ExtractOneVideoFrame::ExtractOneVideoFrame(const DB::FileName &fileName, int offset, QObject *receiver, const char *slot)
+ExtractOneVideoFrame::ExtractOneVideoFrame(const DB::FileName &fileName, double offset, QObject *receiver, const char *slot)
 {
     m_process = new Utilities::Process(this);
     setupWorkingDirectory();
